requirements.txt
pywebview[cef]